check site exists

+++ run.sh
@@ -1,19 +1,25 @@
-export $(egrep -v '^#' .env | xargs)
-bench start&
-export start=$!
-sleep 10
-bench new-site --db-name modehero --db-host db \
-    --mariadb-root-password $MYSQL_ROOT_PASSWORD \
-    --mariadb-root-username root \
-    --admin-password admin \
-    --source_sql /home/frappe/modehero/mysql/backups/modehero.sql \
-    --force modehero.com
+if [ ! -f sites/modehero.com ]; then
 
-sudo kill -9 $start
-sudo pkill -9 redis*
-sudo pkill -9 python*
-bench use modehero.com
-bench start &
-export start=$!
-bench get-app erpnext https://$GITHUB_TOKEN@github.com/modehero/modehero --branch main
-bench --site modehero.com install-app erpnext
+    export $(egrep -v '^#' .env | xargs)
+    bench start&
+    export start=$!
+    sleep 10
+    bench new-site --db-name modehero --db-host db \
+        --mariadb-root-password $MYSQL_ROOT_PASSWORD \
+        --mariadb-root-username root \
+        --admin-password admin \
+        --source_sql /home/frappe/modehero/mysql/backups/modehero.sql \
+        --force modehero.com
+
+    sudo kill -9 $start
+    sudo pkill -9 redis*
+    sudo pkill -9 python*
+    bench use modehero.com
+    bench start &
+    export start=$!
+    bench get-app erpnext https://$GITHUB_TOKEN@github.com/modehero/modehero --branch main
+    bench --site modehero.com install-app erpnext
+
+else
+    bench start
+fi
